% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMetersToGeo.R
\name{addMetersToGeo}
\alias{addMetersToGeo}
\title{addMetersToGeo}
\usage{
addMetersToGeo(d, complete = TRUE)
}
\arguments{
\item{d}{dataframe que tiene latitud y longitud}

\item{complete}{Si es que se dejan solo los registros completos}
}
\value{
dataframe con columnas adicionales en coordenadas cartesianas
}
\description{
Aplica la función coordToMetersDouble a un dataframe y le pega dos columnas con las correspondientes coor
coordenadas cartesianas
}
\examples{
d <- data.frame(lon = -70.57299, lat = -33.39064)
addMetersToGeo(d)
}
