% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordToMetersDouble.R
\name{coordToMetersDouble}
\alias{coordToMetersDouble}
\title{coordToMetersDouble}
\usage{
coordToMetersDouble(vector_lat, vector_lon)
}
\arguments{
\item{vector_lat}{Latitud}

\item{vector_lon}{Longitud}
}
\value{
dataframe con las dos columnas transformadas
}
\description{
Transforma coordenadas geográficas a coordenadas cartesianas
Los valores de está función fueron comprobados en
http://www.csgnetwork.com/degreelenllavcalc.html
}
\examples{
d <- data.frame(lon = -70.57299, lat = -33.39064)
coordToMetersDouble(vector_lon = d$lon, vector_lat = d$lat)
}
