% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{find_functions_with_args}
\alias{find_functions_with_args}
\title{Get all functions which have at least one formal argument matching list of 'arg_names'}
\usage{
find_functions_with_args(package_name, arg_names = NULL)
}
\arguments{
\item{package_name}{string}

\item{arg_names}{character vector of names of arguments to look for. If NULL, then all
functions (and their default formal args are returned).  (default: NULL)}
}
\value{
named list of formal arguments for multiple functions. The name of each
        list element is the function name.
}
\description{
Get all functions which have at least one formal argument matching list of 'arg_names'
}
