% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions.R
\name{init_grrr}
\alias{init_grrr}
\alias{stop_grrr}
\title{Initialize grrr}
\usage{
init_grrr(cutoff = 10, mov = 2, audio = "goku")

stop_grrr()
}
\arguments{
\item{cutoff}{how many errors should occur in a session
before playing an audio file and resetting counter.
Defaults to 10 errors.}

\item{mov}{the absolute distance from the cutoff argument
that the audio file will play (to make the exact count)
a surprise. Defaults to 2.}

\item{audio}{which audio file to play. Default
is a DBZ shout.}
}
\value{
an initialized session with argh
}
\description{
Initialize grrr
}
\examples{
init_grrr()
}
