% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyRut.R
\name{modifyRut}
\alias{modifyRut}
\title{Modifica el rut}
\usage{
modifyRut(x, option = "rut")
}
\arguments{
\item{x}{character que tiene el rut completo}

\item{option}{character que define el output, si es que se quiere obtener el DV o solo el inicio}
}
\value{
character o numeric, dependiendo de las opciones
}
\description{
Sirve para transformar el rut a numérico para mayor velocidad de computo
}
\examples{
modifyRut("0017994104-k", option = "complete")
modifyRut("0017994104-k", option = "dv")
modifyRut("0017994104-k", option = "rut")
}
