% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{sudo_assignInNamespace}
\alias{sudo_assignInNamespace}
\title{A version of 'assignInNamespace' with fewer sanity checks.}
\usage{
sudo_assignInNamespace(x, value, ns, pos = -1,
  envir = as.environment(pos))
}
\arguments{
\item{x}{an object name (given as a character string).}

\item{value}{an \R object.}

\item{ns}{a namespace, or character string giving the namespace.}

\item{pos}{where to look for the object: see \code{\link{get}}.}

\item{envir}{an alternative way to specify an environment to look in.}
}
\description{
A version of 'assignInNamespace' with fewer sanity checks.
}
