% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{update_function_arguments}
\alias{update_function_arguments}
\title{Change the formal arguments of a function}
\usage{
update_function_arguments(function_name, package_name = NULL,
  envir = parent.frame(), ...)
}
\arguments{
\item{function_name}{character name of function}

\item{package_name}{character name of package. Set this if you want the function
to be updated within the package namespace.  If this is unset
then the changed function is placed in the specified environment}

\item{envir}{environment to place the function if package_name not set}

\item{...}{replacement formal arguments for function e.g. "na.rm=FALSE, drop=, n=10"}
}
\value{
TRUE otherwise should throw an error
}
\description{
If the function is specified within a package, then update the function
within the namespace of the package.
}
